* Added simulator * More work on graphics * Fixes for the Tuolaparser

+++ LiveResults.Client/Parsers/TulospalveluParser.cs
@@ -61,23 +61,13 @@ namespace LiveResults.Client.Parsers
             {
                 try
                 {
-
-                    /*var runnerIds = new Dictionary<string, int>();
-                    var bf = new BinaryFormatter();
-                    if (File.Exists("racetimerrunnerids.data"))
+                    var wc = new WebClient
                     {
-                        var st = File.OpenRead("runnerids.data");
-                        runnerIds = bf.Deserialize(st) as Dictionary<string, int>;
-                        st.Close();
-                    }
+                        Encoding = Encoding.GetEncoding("iso-8859-1")
+                    };
 
-
-                    int nextRid = runnerIds.Count > 0 ? runnerIds.Max(x => x.Value) + 1 : 0;*/
-                    var wc = new WebClient();
-                    wc.Encoding = Encoding.GetEncoding("iso-8859-1");
-
-                    Dictionary<string, int> teamTotalTimes = new Dictionary<string, int>();
-                    Dictionary<string, int> teamTotalStatus = new Dictionary<string, int>();
+                    var teamTotalTimes = new Dictionary<string, int>();
+                    var teamTotalStatus = new Dictionary<string, int>();
 
                     var radioControlsRex = new Regex("<a href='/tulokset/en/.*?/tilanne/\\d/(?<control>[1-9])/'>(?<name>.*?)</a>");
                     var rex = new Regex("<h2>(?<classname>.*?)</h2>.*?<table class=\'tuloslista\' cellspacing=0>(?<data>.*?)</table>",
@@ -91,47 +81,36 @@ namespace LiveResults.Client.Parsers
                         {
                             "/"
                         }, StringSplitOptions.RemoveEmptyEntries);
-                        string relayLeg = urlParts[urlParts.Length - 2];
-                        ParseResultList(html, teamTotalTimes, teamTotalStatus, rex, rexBibClub, relayLeg, out className,1000);
 
-                        var radioMatches = radioControlsRex.Matches(html);
-                        foreach (Match rm in radioMatches)
+
+                        bool isSmartView = urlParts[urlParts.Length - 2] == "smart";
+
+                        if (!isSmartView)
                         {
-                            string name = rm.Groups["name"].Value;
-                            string control = rm.Groups["control"].Value;
-                            int code = 1000 + Convert.ToInt32(control);
-                            OnRadioControl(name,code, className, Convert.ToInt32(control));
+                            string relayLeg = null;
+                            if (!isSmartView)
+                                relayLeg = urlParts[urlParts.Length - 2];
 
-                            string rUrl = "http://" + string.Join("/", urlParts.Skip(1).Take(urlParts.Length - 2)) + "/" + control + "/";
-                            html = wc.DownloadString(rUrl);
-                            ParseResultList(html, teamTotalTimes, teamTotalStatus, rex, rexBibClub, relayLeg, out className,code);
-                        }
+                            ParseResultList(html, teamTotalTimes, teamTotalStatus, rex, rexBibClub, relayLeg, out className, 1000);
 
-                        /*foreach (var res in results)
-                        {
-                            var r = idToRunners[res.Key];
-
-                            if (!radios.ContainsKey(r.Class) || radios[r.Class].Count < results[r.ID].Count)
+                            var radioMatches = radioControlsRex.Matches(html);
+                            foreach (Match rm in radioMatches)
                             {
-                                UpdateRadiocontrols(radios, r.Class, results, r.ID);
+                                string name = rm.Groups["name"].Value;
+                                string control = rm.Groups["control"].Value;
+                                int code = 1000 + Convert.ToInt32(control);
+                                OnRadioControl(name, code, className, Convert.ToInt32(control));
+
+                                string rUrl = "http://" + string.Join("/", urlParts.Skip(1).Take(urlParts.Length - 2)) + "/" + control +
+                                              "/";
+                                html = wc.DownloadString(rUrl);
+                                ParseResultList(html, teamTotalTimes, teamTotalStatus, rex, rexBibClub, relayLeg, out className, code);
                             }
-
-                            FireOnResult(new Result
-                            {
-                                ID = res.Key,
-                                Class = r.Class,
-                                RunnerClub = r.Club,
-                                RunnerName = r.Name,
-                                SplitTimes = results[r.ID].OrderBy(x => x.Time).ToList(),
-                                Status = 10,
-                                Time = -10
-
-                            });
-                        }*/
-
-
-
-
+                        }
+                        else
+                        {
+                            ParseSmartViewResultList(html, out className);
+                        }
                     }
 
                     System.Threading.Thread.Sleep(5000);
@@ -198,11 +177,113 @@ namespace LiveResults.Client.Parsers
             }
         }
 
+        private void ParseSmartViewResultList(string html, out string className)
+        {
+            var rex = new Regex("<h2>(?<classname>.*?)</h2>.*?<table class='rva_hk' cellpadding=2 cellspacing=0>(?<data>.*?)</table>",
+                RegexOptions.Singleline);
+            var m = rex.Match(html);
+            var data = m.Groups["data"].Value.Trim();
+            className = m.Groups["classname"].Value.Trim();
+
+            var xml = new XmlDocument();
+            var rexReplace = new Regex("<img border=1 src='.*?' height='11' alt='.*?' title='.*?'>");
+            data = rexReplace.Replace(data, "");
+            data = data.Replace("&nbsp;", " ");
+            xml.LoadXml("<?xml version=\"1.0\" encoding=\"iso-8859-1\"?><table>" + data + "</table>");
+            var trs = xml.GetElementsByTagName("tr");
+
+            var rexRunnerId = new Regex("<a href=.*kilpailijat/(?<id>\\d+)/\\d/\">(?<name>.*?)</a>");
+
+            for (int i = 6; i < trs[0].ChildNodes.Count - 2; i++)
+            {
+                string name = trs[0].ChildNodes[i].InnerXml.Split(new string[] {"<br />"},StringSplitOptions.None)[1];
+                int code = 1001 + i-6;
+                int order = i - 6;
+                OnRadioControl(name, code, className, order);
+            }
+
+            for (int i = 1; i < trs.Count; i++)
+            {
+                string place = trs[i].ChildNodes[0].InnerText.Trim();
+                string name = trs[i].ChildNodes[1].InnerText.Trim();
+                string club = trs[i].ChildNodes[3].InnerText.Trim();
+
+                var ridM = rexRunnerId.Match(trs[i].ChildNodes[1].InnerXml);
+                int id = Convert.ToInt32(ridM.Groups["id"].Value);
+                
+
+                string raceTime = trs[i].ChildNodes[trs[i].ChildNodes.Count-1].InnerText.Trim();
+                int time = -4;
+                int status = 10;
+
+                string startTime = trs[i].ChildNodes[5].InnerText.Trim();
+                DateTime dStartTime = ParseDateTime(startTime);
+                int iStartTime = dStartTime.Hour * 360000 + dStartTime.Minute * 6000 + dStartTime.Second * 100;
+
+                List<ResultStruct> splits = new List<ResultStruct>();
+                for (int split = 6; split < trs[i].ChildNodes.Count - 1; split++)
+                {
+                    var splitCode = 1001 + split - 6;
+                    var splittime = trs[i].ChildNodes[split].InnerText.Trim();
+                    if (!string.IsNullOrEmpty(splittime))
+                    {
+                        splittime = splittime.Substring(0, splittime.IndexOf("(")).Trim();
+                        DateTime dSplitTime = ParseDateTime(splittime);
+                        splits.Add(new ResultStruct()
+                        {
+                            ControlCode = splitCode,
+                            Time = dSplitTime.Hour * 360000 + dSplitTime.Minute * 6000 + dSplitTime.Second * 100
+                        });
+                    }
+                }
+
+                if (raceTime == "DQ" || raceTime == "DNF")
+                {
+                    status = 4;
+                }
+                else if (raceTime == "DNS")
+                {
+                    status = 1;
+                }
+                else if (raceTime.Length > 0)
+                {
+                   /* if (raceTime.StartsWith("+", StringComparison.Ordinal))
+                    {
+                        raceTime = raceTime.Substring(1);
+                        var tid = ParseTimePlus(raceTime);
+                        time = (int) (tid.TotalSeconds * 100);
+                    }
+                    else
+                    {*/
+                        var tid = ParseDateTime(raceTime);
+                        time = tid.Hour * 360000 + tid.Minute * 6000 + tid.Second * 100;
+                    //}
+                    
+
+                    //if (place.Length > 0)
+                    status = 0;
+                }
+
+                FireOnResult(new Result
+                {
+                    ID = id,
+                    Class = className,
+                    RunnerClub = club,
+                    RunnerName = name,
+                    Status = status,
+                    Time = time,
+                    StartTime =  iStartTime,
+                    SplitTimes = splits
+
+                });
+            }
+        }
+
         private void ParseResultList(string html, Dictionary<string, int> teamTotalTimes, Dictionary<string, int> teamTotalStatus, Regex rex, Regex rexBibClub, string relayLeg, out string className, int split)
         {
             var m = rex.Match(html);
             var data = m.Groups["data"].Value.Trim();
-            className = m.Groups["classname"].Value.Trim() + "-" + relayLeg;
+            className = m.Groups["classname"].Value.Trim() + (!string.IsNullOrEmpty(relayLeg) ? "-" + relayLeg : "");
             var xml = new XmlDocument();
             var rexReplace = new Regex("<img border=1 src='.*?' height='11' alt='.*?' title='.*?'>");
             data = rexReplace.Replace(data, "");
@@ -213,7 +294,7 @@ namespace LiveResults.Client.Parsers
             for (int i = 0; i < trs.Count; i++)
             {
                 string place = trs[i].ChildNodes[0].InnerText;
-                string name = trs[i].ChildNodes[5].InnerText;
+                string name = trs[i].ChildNodes[2].InnerText;
                 string club = trs[i].ChildNodes[3].InnerText;
 
                 //var ms = rexBibClub.Match(club);
@@ -221,11 +302,11 @@ namespace LiveResults.Client.Parsers
                 //club = ms.Groups["club"].Value;
 
 
-                string raceTime = trs[i].ChildNodes[4].InnerText.Trim();
+                string raceTime = trs[i].ChildNodes[5].InnerText.Trim();
                 int time = -4;
                 int status = 10;
 
-                int id = Convert.ToInt32(relayLeg) * 1000000 + Convert.ToInt32(bibNo);
+                int id = (!string.IsNullOrEmpty(relayLeg) ?  Convert.ToInt32(relayLeg) * 1000000 : 0) + Convert.ToInt32(bibNo);
 
                 if (raceTime == "DQ" || raceTime == "DNF")
                 {
@@ -257,24 +338,26 @@ namespace LiveResults.Client.Parsers
                         time = bestTime + time;
                     }
 
-                    if (place.Length > 0)
+                    //if (place.Length > 0)
                         status = 0;
                 }
 
-
-                if (relayLeg == "1")
+                if (!string.IsNullOrEmpty(relayLeg))
                 {
-
-                }
-                else
-                {
-                    if (status == 0 && teamTotalStatus[bibNo + "-" + (Convert.ToInt32(relayLeg) - 1)] != 0)
+                    if (relayLeg == "1")
                     {
-                        status = teamTotalStatus[bibNo + "-" + (Convert.ToInt32(relayLeg) - 1)];
+
+                    }
+                    else
+                    {
+                        if (status == 0 && teamTotalStatus[bibNo + "-" + (Convert.ToInt32(relayLeg) - 1)] != 0)
+                        {
+                            status = teamTotalStatus[bibNo + "-" + (Convert.ToInt32(relayLeg) - 1)];
+                        }
                     }
                 }
 
-                string key = bibNo + "-" + relayLeg;
+                string key = bibNo + (!string.IsNullOrEmpty(relayLeg) ? "-" + relayLeg : "");
                 if (!teamTotalTimes.ContainsKey(key))
                     teamTotalTimes.Add(key, time);
                 else
